feat(*): split system configurator

See merge request bmi-digital/dxb!3754

+++ applications/dxb/contentful-migrate/src/migrations/scripts/20220718154551-remove-system-configurator-block-content-type.ts
@@ -1,5 +1,5 @@
 import type { CollectionResponse } from "@bmi-digital/contentful-migration";
-import type { EntryProps } from "contentful-management";
+import type { EntryProps, KeyValueMap } from "contentful-management";
 import type Migration from "contentful-migration";
 import type { MigrationContext, MigrationFunction } from "contentful-migration";
 import {
@@ -8,6 +8,20 @@ import {
 } from "./20210526133509-create-system-configurator-block";
 
 export const description = "Remove System Configurator Block content type";
+
+const delayedRequests = (
+  entries: EntryProps<KeyValueMap>[],
+  cb: (item: EntryProps<KeyValueMap>) => Promise<void>
+) => {
+  let delay = 0;
+  const delayIncrement = 200;
+  return entries.map(async (entry) => {
+    delay += delayIncrement;
+    return new Promise((resolve) => setTimeout(resolve, delay)).then(() =>
+      cb(entry)
+    );
+  });
+};
 
 export const up: MigrationFunction = async (
   migration: Migration,
@@ -18,28 +32,37 @@ export const up: MigrationFunction = async (
     url: `/entries?content_type=systemConfiguratorBlock&limit=1000`
   });
 
-  resp.items.forEach(async (item) => {
-    console.log(`Unpublishing System Configurator Block: ${item.sys.id}`);
+  const unpublishItem = async (entry: EntryProps<KeyValueMap>) => {
+    if (
+      entry.sys.firstPublishedAt &&
+      entry.sys.publishedAt &&
+      entry.sys.publishedCounter
+    ) {
+      console.log(`Unpublishing System Configurator Block: ${entry.sys.id}`);
+      await context!.makeRequest({
+        method: "DELETE",
+        url: `/entries/${entry.sys.id}/published`
+      });
+    }
+  };
+
+  const unpublishedItems = delayedRequests(resp.items, unpublishItem);
+
+  await Promise.all(unpublishedItems);
+
+  const deleteItem = async (entry: EntryProps<KeyValueMap>) => {
+    console.log(`Deleting System Configurator Block: ${entry.sys.id}`);
 
     await context!.makeRequest({
       method: "DELETE",
-      url: `/entries/${item.sys.id}/published`
+      url: `/entries/${entry.sys.id}`
     });
-  });
+  };
 
-  const resp2: CollectionResponse<EntryProps> = await context!.makeRequest({
-    method: "GET",
-    url: `/entries?content_type=systemConfiguratorBlock&limit=1000`
-  });
+  const deletedItems = delayedRequests(resp.items, deleteItem);
 
-  resp2.items.forEach(async (item) => {
-    console.log(`Deleting System Configurator Block: ${item.sys.id}`);
+  await Promise.all(deletedItems);
 
-    await context!.makeRequest({
-      method: "DELETE",
-      url: `/entries/${item.sys.id}`
-    });
-  });
   systemConfiguratorBlockMigrationDown(migration);
 };
